[FLINK-36788] Move GlobalCommitter to flink-runtime
Small refactor for next commit
