+++ flink-runtime/src/main/java/org/apache/flink/streaming/runtime/operators/sink/GlobalCommittableWrapper.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.api.connector.sink2;
+package org.apache.flink.streaming.runtime.operators.sink;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.streaming.runtime.operators.sink.committables.CommittableCollector;
